+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom currency system configuration supported. Base currency and up to 5 custom currencies.
 - Added readied and stowed modifiers to character sheet (under features/tweaks)
 - Fixed capt support department bug
+- Added stress button to character sheet to modify and roll stress, along with tracking breakdowns (page 57 of AWN).
 - Auto show power type on features tab when first added
 - Setting to not roll damage/trauma dice on attack roll automatically with button to roll