+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to the Systems Without Number Redux (SWNR) system for Foundr
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.1] 2025-11-11 CWN Cyberware compendium
+
+- Re-added CWN Cyberware compendium
+
 ## [2.2.0] 2025-11-03 More AWN support
 
 ### Fixes: